refactor: 일부 파일이름의 오타 및 잘못된 이름 수정 및 자바 코드를 기능에 따라 user와 util로 나눠서 분리시킴
