+++ Minecraft/src/main/java/com/example/minecraft/User/SearchUserServlet.java
@@ -1,4 +1,4 @@
-package com.example.minecraft;
+package com.example.minecraft.User;
 
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
