+++ Minecraft/src/main/webapp/userList.jsp
@@ -1,5 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8"
-    pageEncoding="UTF-8" import="java.util.*, com.example.minecraft.*"%>
+    pageEncoding="UTF-8" import="java.util.*, com.example.minecraft.User.*"%>
 <!DOCTYPE html>
 <html>
 <head>
